Adição e correção de comentários + ajustes menores

+++ src/main/java/com/quatrocentosquatro/storemanagement/model/Financeiro.java
@@ -1,23 +1,33 @@
 package com.quatrocentosquatro.storemanagement.model;
-import java.io.Serializable; // Importando Serializable para permitir a serialização da classe
 
+import java.io.Serializable; // Importando Serializable para permitir a serialização da classe
 import java.time.LocalDateTime;
 
-
-// Classe que representa uma transação financeira no sistema.
-// Pode ser uma entrada (receita) ou saída (despesa).
+/**
+ * <p> Classe que representa uma transação financeira no sistema.
+ * <p> Pode ser uma entrada (receita) ou saída (despesa).
+ * 
+ * @author João M. Chervinski
+ */
 public class Financeiro implements Serializable { // Implementa Serializable para permitir a serialização de objetos desta classe
     private static final long serialVersionUID = 1L; // Versão de serialização para compatibilidade
 
     private int id;
     private String tipo; // "entrada" | "saida"
-    private double valor;
+    private float valor;
     private String descricao;
     private LocalDateTime dataHora;
 
-    // Construtor para criar uma nova transação financeira.
-    // Recebe o ID, tipo (entrada/saída), valor e descrição.
-    public Financeiro(int id, String tipo, double valor, String descricao) {
+    /** 
+     * Construtor para criar uma nova transação financeira.
+     * Recebe o ID, tipo (entrada/saída), valor e descrição.
+     * 
+     * @param id        (int)    - O ID da transação.
+     * @param tipo      (String) - O tipo da transação.
+     * @param valor     (float)  - O valor da transação.
+     * @param descricao (String) - A descirção da transação.
+     */
+    public Financeiro(int id, String tipo, float valor, String descricao) {
         this.id = id;
         this.tipo = tipo;
         this.valor = valor;
